Enable building multiple binary packages
With this change we now generate one package for each EFL module, just
as it was done in the 1.7.x series when we had different repositories.
This keeps the compatibility of the existing packages and removes the
need of specifying "Provides" and "Obsoletes" tags.

There is one additional package, efl-data, which ships the localization
files and Efl_Config.h. This file contains only information about the
version.

Change-Id: Iab11c00d07663159d18af24035c8fb7ad6be23dd
Signed-off-by: Philippe Coval <philippe.coval@open.eurogiciel.org>
Signed-off-by: Eduardo Lima (Etrunko) <eduardo.lima@intel.com>