Evas GL: Add hidden flags instead of env vars
merged from http://slp-info.sec.samsung.net/gerrit/#/c/693899/

For Direct Override and Memory Optimization.
This commit includes a lot of changes, but most of them
are directly related to moving the global var into the surface.
In particular, we must keep track of the DR surfaces, so we
add them to a hash indexed by texture id.

Change-Id: I1d584de731f51fbe7a8075983e4116c35b7561bb
Signed-off-by: Daekwang Ryu <dkdk.ryu@samsung.com>