evas/software_generic: Round up merged rect to tile size
Summary:
For some drivers, there is a limitation for partial rects that only multiples of
tile size width/height are allowed.
Currently, rects are merged based on the current rotation, so when the rects
are calculated based on 0 rotation it does not meet the above requirement.
Add a rect rounding up routine after _merge_rects such that the rect is rounded up
to tile size based on 0 rotation.

Change-Id: I23eaccf4a2a263534c686e5d09e2ac62c23698a5