Evas GL: Add hidden flags instead of env vars
merged from http://slp-info.sec.samsung.net/gerrit/#/c/693899/

For Direct Override and Memory Optimization.
This commit includes a lot of changes, but most of them
are directly related to moving the global var into the surface.
In particular, we must keep track of the DR surfaces, so we
add them to a hash indexed by texture id.

Change-Id: I28c131f4c52d09492e1181f0352490b8e0297b90
Signed-off-by: Daekwang Ryu <dkdk.ryu@samsung.com>