Evas GL: Add internal function to create GLES1 specific context
merged form http://slp-info.sec.samsung.net/gerrit/#/c/706789/

This context will be used for GLES 1 indirect rendering.

This context will be used for GLES 1 indirect rendering.
The problem is the following:
eglMakeCurrent( fails with EGL_BAD_MATCH when switching to an
X Pixmap surface for GLES1 indirect rendering because
the EGL configuration for the surface and the context don't match.
The EGLConfig for the context is the sam as that of the window,
while the surface requires the EGL_PIXMAP_BIT flag, as well as other depth,
stencil & msaa flags. (on most drivers, the pixmap bit is not a big deal,
but depth, stencil and msaa are very important).
Thus, the EGLConfig for the surface does't match.
In order to make them match, we need to create another context
to use for GLES1 indirect rendering, that will have the same configuration
as the surface.
In the case of GLES2 we con't have this problem because  GLES2 surfaces
created by Evas GL are not actual EGL surfaces, but simply FBOs and
eglMakeCurrent( is called using the canvas window surface.))

Change-Id: I306cd12308f11dcb2b3f690beb265da1a089b1d5
Signed-off-by: Daekwang Ryu <dkdk.ryu@samsung.com>