Evas GL: Add pixmap surface fallback for gles 1.x indirect rendering
Merge from upstream patch
https://phab.enlightenment.org/rEFLa14492ef73c9267042c5bf505080d0e90af394de

Automatically fallback to indirect rendering on FBO or X11 Pixmap
if the Evas Object Image is not marked as dirty. This should
improve the performance and/or power consumption in those
rare cases where this area of the canvas needs to be redrawn
but the GL content has not changed.

and

Merge from http://slp-info.sec.samsung.net/gerrit/#/c/714550/

- when dirty set is not called but has direct rendering,
   evas gl should render to a fallback surface and use that as texture
 - for gles 1.x, this fallback surface should be a pixmap surface,
   so it uses EVAS_NATIVE_SURFACE_X11 type
 - from now, both EVAS_NATIVE_SURFACE_OPENGL and EVAS_NATIVE_SURFACE_X11
   can be passed for direct rendering

Change-Id: I1cd2592b929b44bd477ae6af60395e36035262b9
Signed-off-by: Mythri Venugopal <mythri.venugopal@samsung.com>