evas-software-generic: Allow future rendering calls if evas has to skip a frame
Summary: If we do not get a new_region_for_update from an engine, it
could mean that the engine has skipped a frame. In that case, we still
want to try and render any future frames. Before this commit, evas
would never try to re-render anything again once it hit a failed
frame. To fix this, we test for a valid returned region, and if we
don't get one, mark this frame as failed so we can try again.

NB: Thanks for the help Cedric !! :)

@fix

Change-Id: I77aa4012483ef8c3474c7509628abd0981317c2e
Signed-off-by: Chris Michael <cp.michael@samsung.com>