ecore_x: switch vsync source as to window visibility.
Enable vsync source drm as default but set timer if the window goes lower or hide.

This behavior avoids using drm vsync as possible.
Since the driver can't fall in sleeping if the process' animator is working.

We know it's problem that animator is working even if app goes background,
but we cannot guranteee the application whether they are using animator or not.

This is better clear way to gurantee drm vsync off.

Change-Id: Ia414b814fd9556a33ae5b2dd74fa15f30a86afda