wayland_shm: check if there is an available buffer before rendering
If all buffers are waiting for buffer release event, it means there are
no available buffers. so render process is cancelled and will never be tried
again if the ecore_evas is not changed any more. This skip of rendering
can cause absence of latest image of ecore_evas. To resolve this unwanted
absence, it's needed to check if there is an available buffer before
rendering and it is also needed to observe release of buffer.

To satisfy those need, Ecore_Evas_Wayland_Shm and "wayland_shm" evas engine is
sharing function pointer and data pointer to communicate in this revision.

Change-Id: I3d2c395d563e09ef7ce50efa69b69ca26a2aa189
Signed-off-by: MinJeong Kim <minjjj.kim@samsung.com>