ecore con: request e19 compositor to create socket
In Tizen 3.0, the security does not allow 1 application to create
socket and other applications to connect to that socket.
All sockets should be created by system daemon or service.
Therefore, we request the e19 compositor to create sockets for us.
Besides, this patches also change the location of sockets to
avoid smack errors.

@tizen_feature

Change-Id: Ia7cb5d937f964d2d7528e84ca06854fcaf805d66