Edje entry: Skip codes for updating cursor when cursor position is not changed
When ever a Edje's cursor function is called, "cursor,changed" signal is
emitted. Even if the position is not changed. And, in Elementary, the signal
will trigger evas_smart_objects_calculate() from elm_widget_show_region_set().
It causes bad performace.

The original patch is being reviewed on EFL upstream.
@tizen_fix

Change-Id: Ib615b7b85fada0ea85dede565bb384288ec50f73