Evas textblock: fix ellipsis item at wrong place in RTL text
Summary:
An ellipsis item could be placed in wrong place in RTL text
with separated and short text items. Normally, an ellipsis
item has to be behind(in logical position) of previous item.
But, when the length of previos text item is 1, ellipsis
item could have same position with the previous item.
It was working fine for LTR text. But, wrong ellipsis could
be shown in RTL text.
@fix

Test Plan:
The test case included.
Run "make check"

Reviewers: tasn, minudf, woohyun, herdsman

Subscribers: Blackmole, z-wony, anand.km, cedric, jpeg

Differential Revision: https://phab.enlightenment.org/D3716

@tizen_fix

Change-Id: I17cee4742ca0e947b2aeb413cf4c1a6b89a15de7